+++ src/main/java/com/codeup/adlister/dao/Planets.java
@@ -9,7 +9,7 @@ public interface Planets {
     // get a list of all the ads
     List<Planet> all();
     // insert a new ad and return the new ad's id
-    Long insert(Planet planet);
+    int insert(Planet planet);
     Planet findPlanetById(int id);
     List<Planet> usersPlanets(int id);
     int updatePlanet(Planet planet);
